Implemented an Abstract class called EnemyShip in the Model:EnemyShips folder.
Changed the collection type of enemyShips to the new Abstract class type EnemeyShip.

Changed bullets from a Collection to an ICollection

Refactored all of the manager classes to where the Canvas is a property in all the Manager classes instead of being a parameter for almost any and all methods.

Moved the collidesWith method in the Bullet class to the GameObject class.

Fixed as many ReSharper warnings I could see while working throughout this commit.