added EnemyShipLevel4 and its corresponding Sprite.
Implemented EnemyShipLevel4 and its corresponding Sprite accordingly.

Changed the names and values of the Rows in the Row enum class.

added a LevelFour value in the ShipLevel enum class.

added a FourthRow value in the Row enum class.

Implemented private helper methods in the EnemyShipManager to help with code smell.

Refactored the values in all of the EnemyShips to match the new enumerated names.

Removed redundant and unused statements in all of the EnemyShipSprites