Added Functionality to where the player has three lives instead of only one.
Added documentation to the EnemyShip abstract class.

Fixed an issue I found while testing the BulletManager. The issue was the bullet wasn't being removed immediately after a collision with a GameObject and therefore was colliding with the player ship multiple times. This resulted in me changing the bullets collection back to a regular Collection instead of ICollection and adding the private helper method RemoveDestroyedBullets.

Added a new delegate and event to the GameManager to detect when the playership was hit.

added a label to the view for the PlayersLives in the bottom right.

Updated the notes.txt file accordingly to relevance.